+++ src/pages/Portfolio/PortfoliosListPage.tsx
@@ -1,0 +1,10 @@
+import BasePage from "@pages/BasePage";
+import styled from "styled-components";
+
+export default function PortfoliosListPage() {
+  return <StyledPortfoliosListPage>hi</StyledPortfoliosListPage>;
+}
+
+const StyledPortfoliosListPage = styled(BasePage)`
+  background-color: #000000;
+`;
