+++ src/router/Routes.ts
@@ -4,8 +4,8 @@ export default {
   SIGNIN: "/signin",
   PROFILE: "/profile",
   DASHBOARD: "/dashboard",
-  WATCHLIST: "/watchlist",
-  PORTFOLIOS: "portfolios",
+  WATCHLISTS: "/watchlists",
+  PORTFOLIOS: "/portfolios",
   PORTFOLIO: "/portfolio/:id",
   PORTFOLIOHOLDING: "/portfolio/:id/holding/:holdingId",
   INDICES: "/indices",
